pyproject.toml
[project]
name = "pilot-gym"
version = "0.1.0"
description = "A reinforcement learning gym environment for autonomous racing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gymnasium>=1.0.0",
    "matplotlib>=3.9.2",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "scipy>=1.14.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    # "UP",
    # flake8-bugbear
    # "B",
    # flake8-simplify
    # "SIM",
    # isort
    "I",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.3",
]
